Fixed the way voterGuidesUpcomingRetrieve was requesting data with 'google_civic_election_id_list'. Moved component OpenExternalWebSite out of utils folder and into Widgets folder. Fixed the way we determine if a voter guide is for a Public Figure vs. Organization.
