+++ src/js/components/VoterGuide/VoterGuideBallot.jsx
@@ -26,7 +26,7 @@ import {
 import ElectionActions from '../../actions/ElectionActions';
 import ElectionStore from '../../stores/ElectionStore';
 import isMobile from '../../utils/isMobile';
-import OpenExternalWebSite from '../../utils/OpenExternalWebSite';
+import OpenExternalWebSite from '../Widgets/OpenExternalWebSite';
 import OrganizationActions from '../../actions/OrganizationActions';
 import OrganizationStore from '../../stores/OrganizationStore';
 import PledgeToSupportOrganizationButton from './PledgeToSupportOrganizationButton';
