+++ src/js/components/Widgets/PositionInformationOnlySnippet.jsx
@@ -4,7 +4,7 @@ import ReactPlayer from 'react-player';
 import { Tooltip, OverlayTrigger } from 'react-bootstrap';
 import { cordovaDot } from '../../utils/cordovaUtils';
 import { renderLog } from '../../utils/logging';
-import OpenExternalWebSite from '../../utils/OpenExternalWebSite';
+import OpenExternalWebSite from './OpenExternalWebSite';
 import ReadMore from './ReadMore';
 import { vimeoRegX, youTubeRegX } from '../../utils/textFormat';
