+++ src/js/routes/Ballot/Measure.jsx
@@ -9,7 +9,7 @@ import MeasureItem from '../../components/Ballot/MeasureItem';
 import MeasureStickyHeader from '../../components/Ballot/MeasureStickyHeader';
 import MeasureActions from '../../actions/MeasureActions';
 import MeasureStore from '../../stores/MeasureStore';
-import OpenExternalWebSite from '../../utils/OpenExternalWebSite';
+import OpenExternalWebSite from '../../components/Widgets/OpenExternalWebSite';
 import OrganizationActions from '../../actions/OrganizationActions';
 import PositionList from '../../components/Ballot/PositionList';
 import SupportActions from '../../actions/SupportActions';
