+++ src/js/routes/VoterGuide/OrganizationVoterGuideCandidate.jsx
@@ -8,7 +8,7 @@ import CandidateStore from '../../stores/CandidateStore';
 import { capitalizeString } from '../../utils/textFormat';
 import GuideList from '../../components/VoterGuide/GuideList';
 import LoadingWheel from '../../components/LoadingWheel';
-import OpenExternalWebSite from '../../utils/OpenExternalWebSite';
+import OpenExternalWebSite from '../../components/Widgets/OpenExternalWebSite';
 import OrganizationActions from '../../actions/OrganizationActions';
 import PositionList from '../../components/Ballot/PositionList';
 import SupportActions from '../../actions/SupportActions';
