+++ src/js/routes/VoterGuide/OrganizationVoterGuideMeasure.jsx
@@ -8,7 +8,7 @@ import { renderLog } from '../../utils/logging';
 import MeasureItem from '../../components/Ballot/MeasureItem';
 import MeasureActions from '../../actions/MeasureActions';
 import MeasureStore from '../../stores/MeasureStore';
-import OpenExternalWebSite from '../../utils/OpenExternalWebSite';
+import OpenExternalWebSite from '../../components/Widgets/OpenExternalWebSite';
 import OrganizationActions from '../../actions/OrganizationActions';
 import PositionList from '../../components/Ballot/PositionList';
 import SupportActions from '../../actions/SupportActions';
